ENH auto memmap large array with loky backend
- CLN correct spelling memmapping
