+++ joblib/pool.py
@@ -13,14 +13,9 @@ that uses a custom alternative to SimpleQueue.
 # Copyright: 2012, Olivier Grisel
 # License: BSD 3 clause
 
-from mmap import mmap
-import errno
 import os
 import stat
 import sys
-import threading
-import atexit
-import tempfile
 import shutil
 import warnings
 from time import sleep
@@ -30,16 +25,6 @@ try:
 except NameError:
     WindowsError = type(None)
 
-from pickle import whichmodule
-try:
-    # Python 2 compat
-    from cPickle import loads
-    from cPickle import dumps
-except ImportError:
-    from pickle import loads
-    from pickle import dumps
-    import copyreg
-
 # Customizable pure Python pickler in Python 2
 # customizable C-optimized pickler under Python 3.3+
 from pickle import Pickler
@@ -48,215 +33,29 @@ from pickle import HIGHEST_PROTOCOL
 from io import BytesIO
 
 from ._multiprocessing_helpers import mp, assert_spawning
+from ._memmapping_reducer import get_memmapping_reducers
+
 # We need the class definition to derive from it not the multiprocessing.Pool
 # factory function
 from multiprocessing.pool import Pool
 
 try:
     import numpy as np
-    from numpy.lib.stride_tricks import as_strided
 except ImportError:
     np = None
 
-from .numpy_pickle import load
-from .numpy_pickle import dump
-from .hashing import hash
-from .backports import make_memmap
+if sys.version_info[:2] > (2, 7):
+    import copyreg
+
 # Some system have a ramdisk mounted by default, we can use it instead of /tmp
 # as the default folder to dump big arrays to share with subprocesses
 SYSTEM_SHARED_MEM_FS = '/dev/shm'
 
 # Folder and file permissions to chmod temporary files generated by the
-# memmaping pool. Only the owner of the Python process can access the
+# memmapping pool. Only the owner of the Python process can access the
 # temporary files and folder.
 FOLDER_PERMISSIONS = stat.S_IRUSR | stat.S_IWUSR | stat.S_IXUSR
 FILE_PERMISSIONS = stat.S_IRUSR | stat.S_IWUSR
-
-###############################################################################
-# Support for efficient transient pickling of numpy data structures
-
-
-def _get_backing_memmap(a):
-    """Recursively look up the original np.memmap instance base if any."""
-    b = getattr(a, 'base', None)
-    if b is None:
-        # TODO: check scipy sparse datastructure if scipy is installed
-        # a nor its descendants do not have a memmap base
-        return None
-
-    elif isinstance(b, mmap):
-        # a is already a real memmap instance.
-        return a
-
-    else:
-        # Recursive exploration of the base ancestry
-        return _get_backing_memmap(b)
-
-
-def has_shareable_memory(a):
-    """Return True if a is backed by some mmap buffer directly or not."""
-    return _get_backing_memmap(a) is not None
-
-
-def _strided_from_memmap(filename, dtype, mode, offset, order, shape, strides,
-                         total_buffer_len):
-    """Reconstruct an array view on a memory mapped file."""
-    if mode == 'w+':
-        # Do not zero the original data when unpickling
-        mode = 'r+'
-
-    if strides is None:
-        # Simple, contiguous memmap
-        return make_memmap(filename, dtype=dtype, shape=shape, mode=mode,
-                           offset=offset, order=order)
-    else:
-        # For non-contiguous data, memmap the total enclosing buffer and then
-        # extract the non-contiguous view with the stride-tricks API
-        base = make_memmap(filename, dtype=dtype, shape=total_buffer_len,
-                           mode=mode, offset=offset, order=order)
-        return as_strided(base, shape=shape, strides=strides)
-
-
-def _reduce_memmap_backed(a, m):
-    """Pickling reduction for memmap backed arrays.
-
-    a is expected to be an instance of np.ndarray (or np.memmap)
-    m is expected to be an instance of np.memmap on the top of the ``base``
-    attribute ancestry of a. ``m.base`` should be the real python mmap object.
-    """
-    # offset that comes from the striding differences between a and m
-    a_start, a_end = np.byte_bounds(a)
-    m_start = np.byte_bounds(m)[0]
-    offset = a_start - m_start
-
-    # offset from the backing memmap
-    offset += m.offset
-
-    if m.flags['F_CONTIGUOUS']:
-        order = 'F'
-    else:
-        # The backing memmap buffer is necessarily contiguous hence C if not
-        # Fortran
-        order = 'C'
-
-    if a.flags['F_CONTIGUOUS'] or a.flags['C_CONTIGUOUS']:
-        # If the array is a contiguous view, no need to pass the strides
-        strides = None
-        total_buffer_len = None
-    else:
-        # Compute the total number of items to map from which the strided
-        # view will be extracted.
-        strides = a.strides
-        total_buffer_len = (a_end - a_start) // a.itemsize
-    return (_strided_from_memmap,
-            (m.filename, a.dtype, m.mode, offset, order, a.shape, strides,
-             total_buffer_len))
-
-
-def reduce_memmap(a):
-    """Pickle the descriptors of a memmap instance to reopen on same file."""
-    m = _get_backing_memmap(a)
-    if m is not None:
-        # m is a real mmap backed memmap instance, reduce a preserving striding
-        # information
-        return _reduce_memmap_backed(a, m)
-    else:
-        # This memmap instance is actually backed by a regular in-memory
-        # buffer: this can happen when using binary operators on numpy.memmap
-        # instances
-        return (loads, (dumps(np.asarray(a), protocol=HIGHEST_PROTOCOL),))
-
-
-class ArrayMemmapReducer(object):
-    """Reducer callable to dump large arrays to memmap files.
-
-    Parameters
-    ----------
-    max_nbytes: int
-        Threshold to trigger memmaping of large arrays to files created
-        a folder.
-    temp_folder: str
-        Path of a folder where files for backing memmaped arrays are created.
-    mmap_mode: 'r', 'r+' or 'c'
-        Mode for the created memmap datastructure. See the documentation of
-        numpy.memmap for more details. Note: 'w+' is coerced to 'r+'
-        automatically to avoid zeroing the data on unpickling.
-    verbose: int, optional, 0 by default
-        If verbose > 0, memmap creations are logged.
-        If verbose > 1, both memmap creations, reuse and array pickling are
-        logged.
-    prewarm: bool, optional, False by default.
-        Force a read on newly memmaped array to make sure that OS pre-cache it
-        memory. This can be useful to avoid concurrent disk access when the
-        same data array is passed to different worker processes.
-    """
-
-    def __init__(self, max_nbytes, temp_folder, mmap_mode, verbose=0,
-                 context_id=None, prewarm=True):
-        self._max_nbytes = max_nbytes
-        self._temp_folder = temp_folder
-        self._mmap_mode = mmap_mode
-        self.verbose = int(verbose)
-        self._prewarm = prewarm
-        if context_id is not None:
-            warnings.warn('context_id is deprecated and ignored in joblib'
-                          ' 0.9.4 and will be removed in 0.11',
-                          DeprecationWarning)
-
-    def __call__(self, a):
-        m = _get_backing_memmap(a)
-        if m is not None:
-            # a is already backed by a memmap file, let's reuse it directly
-            return _reduce_memmap_backed(a, m)
-
-        if (not a.dtype.hasobject
-                and self._max_nbytes is not None
-                and a.nbytes > self._max_nbytes):
-            # check that the folder exists (lazily create the pool temp folder
-            # if required)
-            try:
-                os.makedirs(self._temp_folder)
-                os.chmod(self._temp_folder, FOLDER_PERMISSIONS)
-            except OSError as e:
-                if e.errno != errno.EEXIST:
-                    raise e
-
-            # Find a unique, concurrent safe filename for writing the
-            # content of this array only once.
-            basename = "%d-%d-%s.pkl" % (
-                os.getpid(), id(threading.current_thread()), hash(a))
-            filename = os.path.join(self._temp_folder, basename)
-
-            # In case the same array with the same content is passed several
-            # times to the pool subprocess children, serialize it only once
-
-            # XXX: implement an explicit reference counting scheme to make it
-            # possible to delete temporary files as soon as the workers are
-            # done processing this data.
-            if not os.path.exists(filename):
-                if self.verbose > 0:
-                    print("Memmaping (shape=%r, dtype=%s) to new file %s" % (
-                        a.shape, a.dtype, filename))
-                for dumped_filename in dump(a, filename):
-                    os.chmod(dumped_filename, FILE_PERMISSIONS)
-
-                if self._prewarm:
-                    # Warm up the data to avoid concurrent disk access in
-                    # multiple children processes
-                    load(filename, mmap_mode=self._mmap_mode).max()
-            elif self.verbose > 1:
-                print("Memmaping (shape=%s, dtype=%s) to old file %s" % (
-                    a.shape, a.dtype, filename))
-
-            # The worker process will use joblib.load to memmap the data
-            return (load, (filename, self._mmap_mode))
-        else:
-            # do not convert a into memmap, let pickler do its usual copy with
-            # the default system pickler
-            if self.verbose > 1:
-                print("Pickling array (shape=%r, dtype=%s)." % (
-                    a.shape, a.dtype))
-            return (loads, (dumps(a, protocol=HIGHEST_PROTOCOL),))
 
 
 ###############################################################################
@@ -438,7 +237,7 @@ def delete_folder(folder_path):
         warnings.warn("Failed to clean temporary folder: %s" % folder_path)
 
 
-class MemmapingPool(PicklingPool):
+class MemmappingPool(PicklingPool):
     """Process pool that shares large arrays to avoid memory copy.
 
     This drop-in replacement for `multiprocessing.pool.Pool` makes
@@ -453,7 +252,7 @@ class MemmapingPool(PicklingPool):
 
     Furthermore large arrays from the parent process are automatically
     dumped to a temporary folder on the filesystem such as child
-    processes to access their content via memmaping (file system
+    processes to access their content via memmapping (file system
     backed shared memory).
 
     Note: it is important to call the terminate method to collect
@@ -468,7 +267,7 @@ class MemmapingPool(PicklingPool):
     initargs: tuple, optional
         Arguments passed to the initializer callable.
     temp_folder: str, optional
-        Folder to be used by the pool for memmaping large arrays
+        Folder to be used by the pool for memmapping large arrays
         for sharing memory with worker processes. If None, this will try in
         order:
         - a folder pointed by the JOBLIB_TEMP_FOLDER environment variable,
@@ -480,7 +279,7 @@ class MemmapingPool(PicklingPool):
     max_nbytes int or None, optional, 1e6 by default
         Threshold on the size of arrays passed to the workers that
         triggers automated memory mapping in temp_folder.
-        Use None to disable memmaping of large arrays.
+        Use None to disable memmapping of large arrays.
     mmap_mode: {'r+', 'r', 'w+', 'c'}
         Memmapping mode for numpy arrays passed to workers.
         See 'max_nbytes' parameter documentation for more details.
@@ -492,13 +291,14 @@ class MemmapingPool(PicklingPool):
         master process.
     verbose: int, optional
         Make it possible to monitor how the communication of numpy arrays
-        with the subprocess is handled (pickling or memmaping)
+        with the subprocess is handled (pickling or memmapping)
     prewarm: bool or str, optional, "auto" by default.
-        If True, force a read on newly memmaped array to make sure that OS pre-
-        cache it in memory. This can be useful to avoid concurrent disk access
-        when the same data array is passed to different worker processes.
-        If "auto" (by default), prewarm is set to True, unless the Linux shared
-        memory partition /dev/shm is available and used as temp_folder.
+        If True, force a read on newly memmapped array to make sure that OS
+        pre-cache it in memory. This can be useful to avoid concurrent disk
+        access when the same data array is passed to different worker
+        processes. If "auto" (by default), prewarm is set to True, unless the
+        Linux shared memory partition /dev/shm is available and used as temp
+        folder.
 
     `forward_reducers` and `backward_reducers` are expected to be
     dictionaries with key/values being `(type, callable)` pairs where
@@ -513,97 +313,30 @@ class MemmapingPool(PicklingPool):
     def __init__(self, processes=None, temp_folder=None, max_nbytes=1e6,
                  mmap_mode='r', forward_reducers=None, backward_reducers=None,
                  verbose=0, context_id=None, prewarm=False, **kwargs):
-        if forward_reducers is None:
-            forward_reducers = dict()
-        if backward_reducers is None:
-            backward_reducers = dict()
+
         if context_id is not None:
             warnings.warn('context_id is deprecated and ignored in joblib'
                           ' 0.9.4 and will be removed in 0.11',
                           DeprecationWarning)
-
-        # Prepare a sub-folder name for the serialization of this particular
-        # pool instance (do not create in advance to spare FS write access if
-        # no array is to be dumped):
-        use_shared_mem = False
-        pool_folder_name = "joblib_memmaping_pool_%d_%d" % (
-            os.getpid(), id(self))
-        if temp_folder is None:
-            temp_folder = os.environ.get('JOBLIB_TEMP_FOLDER', None)
-        if temp_folder is None:
-            if os.path.exists(SYSTEM_SHARED_MEM_FS):
-                try:
-                    temp_folder = SYSTEM_SHARED_MEM_FS
-                    pool_folder = os.path.join(temp_folder, pool_folder_name)
-                    if not os.path.exists(pool_folder):
-                        os.makedirs(pool_folder)
-                    use_shared_mem = True
-                except IOError:
-                    # Missing rights in the the /dev/shm partition,
-                    # fallback to regular temp folder.
-                    temp_folder = None
-        if temp_folder is None:
-            # Fallback to the default tmp folder, typically /tmp
-            temp_folder = tempfile.gettempdir()
-        temp_folder = os.path.abspath(os.path.expanduser(temp_folder))
-        pool_folder = os.path.join(temp_folder, pool_folder_name)
-        self._temp_folder = pool_folder
-
-        # Register the garbage collector at program exit in case caller forgets
-        # to call terminate explicitly: note we do not pass any reference to
-        # self to ensure that this callback won't prevent garbage collection of
-        # the pool instance and related file handler resources such as POSIX
-        # semaphores and pipes
-        pool_module_name = whichmodule(delete_folder, 'delete_folder')
-
-        def _cleanup():
-            # In some cases the Python runtime seems to set delete_folder to
-            # None just before exiting when accessing the delete_folder
-            # function from the closure namespace. So instead we reimport
-            # the delete_folder function explicitly.
-            # https://github.com/joblib/joblib/issues/328
-            # We cannot just use from 'joblib.pool import delete_folder'
-            # because joblib should only use relative imports to allow
-            # easy vendoring.
-            delete_folder = __import__(
-                pool_module_name, fromlist=['delete_folder']).delete_folder
-            delete_folder(pool_folder)
-
-        atexit.register(_cleanup)
-
-        if np is not None:
-            # Register smart numpy.ndarray reducers that detects memmap backed
-            # arrays and that is alse able to dump to memmap large in-memory
-            # arrays over the max_nbytes threshold
-            if prewarm == "auto":
-                prewarm = not use_shared_mem
-            forward_reduce_ndarray = ArrayMemmapReducer(
-                max_nbytes, pool_folder, mmap_mode, verbose,
+        forward_reducers, backward_reducers, self._temp_folder = \
+            get_memmapping_reducers(
+                id(self), temp_folder=temp_folder, max_nbytes=max_nbytes,
+                mmap_mode=mmap_mode, forward_reducers=forward_reducers,
+                backward_reducers=backward_reducers, verbose=verbose,
                 prewarm=prewarm)
-            forward_reducers[np.ndarray] = forward_reduce_ndarray
-            forward_reducers[np.memmap] = reduce_memmap
-
-            # Communication from child process to the parent process always
-            # pickles in-memory numpy.ndarray without dumping them as memmap
-            # to avoid confusing the caller and make it tricky to collect the
-            # temporary folder
-            backward_reduce_ndarray = ArrayMemmapReducer(
-                None, pool_folder, mmap_mode, verbose)
-            backward_reducers[np.ndarray] = backward_reduce_ndarray
-            backward_reducers[np.memmap] = reduce_memmap
 
         poolargs = dict(
             processes=processes,
             forward_reducers=forward_reducers,
             backward_reducers=backward_reducers)
         poolargs.update(kwargs)
-        super(MemmapingPool, self).__init__(**poolargs)
+        super(MemmappingPool, self).__init__(**poolargs)
 
     def terminate(self):
         n_retries = 10
         for i in range(n_retries):
             try:
-                super(MemmapingPool, self).terminate()
+                super(MemmappingPool, self).terminate()
                 break
             except OSError as e:
                 if isinstance(e, WindowsError):
